Improve single work edit form and change video thumbnails
Allow editing of strings and URIs in the same field. When loading authority fields for this form, display labels instead of URIs. Normalize spacing in works_controller_behavior.rb.

Customize Hydra Derivatives to create a video thumbnail from midway through the video rather than 2 seconds in. This is to avoid black thumbnails in videos that are blank for the first few seconds.