Explicitly add admins to edit groups for newly created private works.
Previously, the "total viewable items" didn't include admins in the edit group even if admins could already edit private works deposited by others. This caused a mismatch between that and the true number of viewable items. This includes admins explicitly if a work is set to private so that the numbers match.