Remove "public/uv" from gitignore and add public/uv files.
Hyrax's approach is to install via rake task/yarn:
  1. creates a directory node_modules in Rails root (not tracked by git)
  2. copy files from node_modules/universalviewer/dist to public/uv (also not tracked)
  3. copy config files

Our approach in docker:
  1. install universalviewer via npm into /node/node_modules (see Dockerfile)
  2. track all files in public/uv because we have modified public/uv/lib/GalleryComponent.js, public/uv/uv.html, and added other custom files

This way, a docker app will work immediately after docker-compose build and customized files are tracked and consistent across installs. Performance *might* be better, since the entire /data dir is a docker volume so we prefer to have fewer files in /data if possible. This is at the cost of tracking more files in public/uv that will have to change when uv is eventually upgraded. Also, any change to how uv is installed (npm, yarn) will require modifying the Dockerfile.