Add rescue block to AttachFilesToWorkJob and make it idempotent.
This job fails intermittently with a "Stack level too deep error" when a work has >200 file sets.

Previously, the job would not reattach the file because of a line that checks for the presence of a file set URI and returns if true. However, an uploaded file can still have a file set URI as long as the IngestJob succeeds, even if that file set isn't actually attached to anything.

I've amended the job to check for a parent/child relationship before skipping and to reattach the file if the file set has no parent.