Do not create thumbnails since we use IIIF for this instead.
We need bigger thumbnails than the ones created with Hydra Derivatives (e.g. on the homepage) or the thumbnail for a file's latest version, so thumbs created with HD are of little use.

Also, there is a shell script that we can modify to delete these cached thumbnails if file storage becomes an issue (see Riiif gem README). No such script exists for derivatives generated with HD.