VAULT-218 Reindex a collection's nested members if the collection's title or nesting relationships change(s).
See also commit 7254761c9add95afc6e2a60c8c700cfdb8aadb54. Note that we need to add a title to all collection params in controller specs. This is ok since a collection must have a title to be created, and therefore we can always expect a title in the params.